log UserPersonalization api request method

+++ api/UserPersonalization.cs
@@ -43,6 +43,8 @@ namespace Editor
 
             string method = req.Method.ToLowerInvariant();
 
+            log.LogInformation($"UserPersonalization request {method}");
+
             if (method == "get")
             {
                 return await GetUserPersonalization(table, userId);
@@ -51,10 +53,10 @@ namespace Editor
             {
                 return await SaveUserPersonalization(req, table, userId);
             }
-            /*else if (method == "delete")
+            else if (method == "delete")
             {
                 return await DeleteUserPersonalization(table, userId);
-            }*/
+            }
             else
             {
                 return new BadRequestObjectResult($"Method '{req.Method}' not supported");
